Update espresso test
